Created data:shared, domain:shared, core:kotlin modules with base code for models el7mdullah & inshallah
