+++ .idea/compiler.xml
@@ -10,6 +10,10 @@
       <module name="Simple_NYTimes_Showcase.feature.articles.domains" target="11" />
       <module name="Simple_NYTimes_Showcase.feature.articles.domains.main" target="11" />
       <module name="Simple_NYTimes_Showcase.feature.articles.domains.test" target="11" />
+      <module name="SimpleNYTimesShowcase.feature.articles.data" target="21" />
+      <module name="SimpleNYTimesShowcase.feature.articles.domain" target="11" />
+      <module name="SimpleNYTimesShowcase.feature.articles.domain.main" target="11" />
+      <module name="SimpleNYTimesShowcase.feature.articles.domain.test" target="11" />
     </bytecodeTargetLevel>
   </component>
 </project>
