+++ .idea/gradle.xml
@@ -11,10 +11,14 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/feature" />
-            <option value="$PROJECT_DIR$/feature/articles" />
-            <option value="$PROJECT_DIR$/feature/articles/data" />
-            <option value="$PROJECT_DIR$/feature/articles/domain" />
+            <option value="$PROJECT_DIR$/core" />
+            <option value="$PROJECT_DIR$/core/kotlin" />
+            <option value="$PROJECT_DIR$/data" />
+            <option value="$PROJECT_DIR$/data/articles" />
+            <option value="$PROJECT_DIR$/data/shared" />
+            <option value="$PROJECT_DIR$/domain" />
+            <option value="$PROJECT_DIR$/domain/articles" />
+            <option value="$PROJECT_DIR$/domain/shared" />
           </set>
         </option>
       </GradleProjectSettings>
