+++ core/kotlin/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+	id("java-library")
+	alias(libs.plugins.jetbrains.kotlin.jvm)
+}
+java {
+	sourceCompatibility = JavaVersion.VERSION_11
+	targetCompatibility = JavaVersion.VERSION_11
+}
+kotlin {
+	compilerOptions {
+		jvmTarget = org.jetbrains.kotlin.gradle.dsl.JvmTarget.JVM_11
+	}
+}
