+++ data/articles/src/main/AndroidManifest.xml
