+++ data/articles/src/main/java/my/ym/feature_articles_data/models/ApiMediaMetadata.kt
