+++ data/articles/src/main/java/my/ym/feature_articles_data/models/utils/ApiViewedArticleMapper.kt
@@ -1,0 +1,21 @@
+package my.ym.feature_articles_data.models.utils
+
+import my.ym.data_shared.models.ApiDateAndTimeFormat
+import my.ym.data_shared.models.ApiDateFormat
+import my.ym.feature_articles_data.models.ApiViewedArticle
+import my.ym.feature_articles_domain.models.AppViewedArticle
+
+private const val SEMI_COLON = ';'
+
+fun ApiViewedArticle.toAppViewedArticle(): AppViewedArticle {
+	return AppViewedArticle(
+		title = title.orEmpty(),
+		summary = summary.orEmpty(),
+		section = section.orEmpty(),
+		byLine = byLine.orEmpty(),
+		publishedAt = ApiDateFormat(publishedDate).toLocalDateOrNull(),
+		lastUpdatedAt = ApiDateAndTimeFormat(lastUpdatedDateAndTime).toLocalDateTimeOrNull(),
+		keywords = keywords.orEmpty().split(SEMI_COLON),
+		media = media.orEmpty().map { it.toAppMedia() }
+	)
+}
