+++ domain/articles/src/main/java/my/ym/feature_articles_domain/models/AppViewedArticle.kt
