--- feature/articles/data/build.gradle.kts
@@ -1,59 +1,0 @@
-plugins {
-	alias(libs.plugins.android.library)
-	alias(libs.plugins.kotlin.android)
-}
-
-android {
-	namespace = "my.ym.feature_articles_data"
-	compileSdk = 35
-
-	defaultConfig {
-		minSdk = 24
-
-		testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-		consumerProguardFiles("consumer-rules.pro")
-	}
-
-	buildTypes {
-		release {
-			isMinifyEnabled = false
-			proguardFiles(
-				getDefaultProguardFile("proguard-android-optimize.txt"),
-				"proguard-rules.pro"
-			)
-		}
-	}
-	compileOptions {
-		// Java 8+ API desugaring support
-		isCoreLibraryDesugaringEnabled = true
-
-		sourceCompatibility = JavaVersion.VERSION_11
-		targetCompatibility = JavaVersion.VERSION_11
-	}
-	kotlinOptions {
-		jvmTarget = "11"
-	}
-}
-
-dependencies {
-	// Java 8+ API desugaring support
-	coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:2.0.3")
-
-	// Local Modules
-	implementation(projects.feature.articles.domain)
-
-	// Androidx
-	implementation(libs.androidx.core.ktx)
-
-	// Retrofit
-	implementation(libs.retrofit2.retrofit)
-	implementation(libs.retrofit2.converter.gson)
-
-	// Test
-	testImplementation(libs.junit)
-
-	// Test Android
-	androidTestImplementation(libs.androidx.junit)
-	androidTestImplementation(libs.androidx.espresso.core)
-
-}
